+++ runtime/browser/android/xwalk_settings.h
@@ -18,7 +18,7 @@ class XWalkRenderViewHostExt;
 
 class XWalkSettings : public content::WebContentsObserver {
  public:
-  XWalkSettings(JNIEnv* env, jobject obj, jint web_contents);
+  XWalkSettings(JNIEnv* env, jobject obj, jlong web_contents);
   virtual ~XWalkSettings();
 
   // Called from Java.
